+++ src/index.ts
@@ -8,8 +8,6 @@
  * @module webkms
  */
 export { AsymmetricKey } from './AsymmetricKey.js'
-export { CapabilityAgent } from './CapabilityAgent.js'
-export type { VerificationKeyDescriptor } from './CapabilityAgent.js'
 export { Hmac } from './Hmac.js'
 export { Kek } from './Kek.js'
 export { KeyAgreementKey } from './KeyAgreementKey.js'